% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR plots.R
\name{generic_line}
\alias{generic_line}
\title{Generic line plot}
\usage{
generic_line(x, y, group = NULL, panel = NULL, data, col = TRUE)
}
\arguments{
\item{x}{x variable. Can be continuous numeric variable (e.g. year) or character (e.g. financial year)}

\item{y}{y variable. A numeric variable (e.g. catch)}

\item{group}{A grouping variable (e.g. fishery)}

\item{panel}{A panelling variable (e.g. species)}

\item{data}{A \code{data.frame} containing at least columns \code{x} and \code{y} and optionally \code{group} and \code{panel}}

\item{col}{Should the graph be in colour? (TRUE or FALSE)}
}
\value{
A ggplot object
}
\description{
A generic bivariate line plot. Minimally requires x and y variables but can also take additional grouping and paneling variables.
}
\examples{
generic_line(year, catch, fishery, species, data=data)
}

