% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR plots.R
\name{species_bar}
\alias{species_bar}
\title{Species catch barplot}
\usage{
species_bar(data, Species, fishery_code, fishery_names,
  start_year = min(data$year, na.rm = T), end_year = max(data$year, na.rm =
  T), col = TRUE, save = FALSE, ...)
}
\arguments{
\item{data}{A data frame of a CAES data extraction}

\item{Species}{Species name, exactly as per `FishCubename` e.g. \code{"Red Emperor"}. May also be a vector e.g. \code{c("Red Emperor", "Goldband Snapper")}}

\item{fishery_code}{Fishery codes used by CAES e.g \code{"NDS"}. May also be a vector e.g. \code{c("NDS","KTR")}}

\item{fishery_names}{Desired fishery name(s). Must be same length as \code{fishery_code} e.g. \code{c("Northern Demersal","Northern Demersal")}}

\item{start_year}{}

\item{end_year}{}

\item{col}{Should the graph be in colour? (TRUE or FALSE)}

\item{save}{Saves a PNG file (TRUE or FALSE) using the size of the current graphics device (i.e. may not work very well)}
}
\value{
A ggplot object
}
\description{
Generates a barplot of species catch using data in the format of a standard CAES extraction.
}

