% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR plots.R
\name{species_barplot}
\alias{species_barplot}
\title{Barplot of species catch for multiple fisheries}
\usage{
species_barplot(data, Species, fishery_code, fishery_names,
  start_year = min(data$year, na.rm = T), end_year = max(data$year, na.rm =
  T), save = FALSE, ...)
}
\arguments{
\item{data}{A data frame of a CAES data extraction}

\item{Species}{Species name, exactly as per `FishCubename` e.g. "Red Emperor". May also be a vector e.g. c("Red Emperor", "Goldband Snapper")}

\item{fishery_code}{Fishery codes used by CAES e.g "NDS". May also be a vector e.g. c("NDS","KTR")}

\item{fishery_names}{Desired fishery name(s). Must be same length as \code{fishery_code} e.g. c("Northern Demersal","Northern Demersal")}

\item{start_year}{}

\item{end_year}{}

\item{save}{Saves a PNG file (TRUE or FALSE)}
}
\value{
A ggplot object
}
\description{
Barplot of species catch for multiple fisheries
}

