% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuna}
\alias{tuna}
\title{Global catch of yellowfin and bigeye tunas 1950 - 1980}
\format{A data frame with 488 rows and 4 variables}
\source{
\url{http://www.fao.org/}
}
\usage{
tuna
}
\description{
Nominal annual catch of yellowfin and bigeye tuna reported by FAO
}
\keyword{datasets}

