% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delta_mean}
\alias{delta_mean}
\title{Delta log normal mean}
\usage{
delta_mean(x, na.rm = FALSE, log = FALSE, CI = FALSE)
}
\arguments{
\item{x}{numeric. The variable of interest.}

\item{na.rm}{logical. Should NA values be removed?}

\item{log}{logical. If FALSE, the exponent of the mean and confidence intervals is returned}

\item{CI}{logical. If true returns the standard deviation and upper and lower 95 per cent confidence intervals of the mean}
}
\value{
mean The mean of \code{x} adjusted for zero-inflation and lognormal bias

sd The standard deviation of \code{x} (square root of the variance) adjusted for zero-inflation and lognormal bias

p Zero inflation factor. The proportion of \code{x}>0

lower,upper The upper and lower 95 per cent confidence intervals of the mean
}
\description{
Function for calculating the mean of a zero-inflated log normal variable using the delta method. 
Implements the Cox method in Fletcher (2008) for calculating the mean and approximate
confidence interval for a zero-inflated lognormal variable, or log normal variable when all values 
are >0
}
\examples{
data(sunspots) 
delta_mean(sunspots,CI=TRUE)
}
\references{
Fletcher, D. (2008) Confidence intervals for the mean of the delta-lognormal distribution. \emph{Environ Ecol Stat} 15: 175-189
}

